removed tester flag and replaced old with new components / fixed a minor header icon padding bug
