+++ components/Frame/Header.js
@@ -1,12 +1,7 @@
-import React, { Component, Fragment } from 'react'
-import { css, merge } from 'glamor'
+import React, { useMemo, useState, useRef, useEffect } from 'react'
+import { css } from 'glamor'
 import { compose } from 'react-apollo'
 import { withRouter } from 'next/router'
-
-import withT from '../../lib/withT'
-import withInNativeApp, { postMessage } from '../../lib/withInNativeApp'
-import { Router, matchPath } from '../../lib/routes'
-
 import {
   Logo,
   colors,
@@ -15,382 +10,194 @@ import {
   HeaderHeightProvider
 } from '@project-r/styleguide'
 
+import { Router } from '../../lib/routes'
 import { withMembership } from '../Auth/checkRoles'
+import withT from '../../lib/withT'
+import withInNativeApp, { postMessage } from '../../lib/withInNativeApp'
+import { shouldIgnoreClick } from '../../lib/utils/link'
+import NotificationIcon from '../Notifications/NotificationIcon'
+import BackIcon from '../Icons/Back'
+import HLine from '../Frame/HLine'
 
-import Toggle from './Toggle'
 import User from './User'
 import Popover from './Popover'
 import NavPopover from './Popover/Nav'
+import UserNavPopover from './Popover/UserNav'
 import LoadingBar from './LoadingBar'
 import Pullable from './Pullable'
-
-import { MdSearch } from 'react-icons/md'
-import BackIcon from '../Icons/Back'
-
-import { shouldIgnoreClick } from '../../lib/utils/link'
+import Toggle from './Toggle'
+import SecondaryNav from './SecondaryNav'
 
 import {
   HEADER_HEIGHT,
   HEADER_HEIGHT_MOBILE,
-  HEADER_ICON_SIZE,
-  ZINDEX_HEADER,
+  SUBHEADER_HEIGHT,
+  ZINDEX_POPOVER,
   LOGO_WIDTH,
   LOGO_PADDING,
   LOGO_WIDTH_MOBILE,
   LOGO_PADDING_MOBILE,
-  HEADER_HEIGHT_CONFIG
+  TRANSITION_MS
 } from '../constants'
-
-import HeaderIconA from './HeaderIconA'
-
-import NotificationIcon from '../Notifications/NotificationIcon'
-
-const TRANSITION_MS = 200
-
-const styles = {
-  bar: css({
-    zIndex: ZINDEX_HEADER,
-    position: 'fixed',
-    '@media print': {
-      position: 'absolute'
-    },
-    top: 0,
-    left: 0,
-    right: 0
-  }),
-  barOpaque: css({
-    height: HEADER_HEIGHT_MOBILE,
-    [mediaQueries.mUp]: {
-      height: HEADER_HEIGHT
-    },
-    '@media print': {
-      backgroundColor: 'transparent'
-    }
-  }),
-  center: css({
-    margin: '0 auto 0',
-    padding: '0 60px',
-    textAlign: 'center',
-    transition: `opacity ${TRANSITION_MS}ms ease-in-out`
-  }),
-  logo: css({
-    position: 'relative',
-    display: 'inline-block',
-    padding: LOGO_PADDING_MOBILE,
-    width: LOGO_WIDTH_MOBILE + LOGO_PADDING_MOBILE * 2,
-    [mediaQueries.mUp]: {
-      padding: LOGO_PADDING,
-      width: LOGO_WIDTH + LOGO_PADDING * 2
-    },
-    verticalAlign: 'middle'
-  }),
-  leftItem: css({
-    '@media print': {
-      display: 'none'
-    },
-    transition: `opacity ${TRANSITION_MS}ms ease-in-out`
-  }),
-  back: css({
-    display: 'block',
-    position: 'absolute',
-    left: 0,
-    top: -1,
-    padding: '10px 10px 10px 15px',
-    [mediaQueries.mUp]: {
-      top: -1 + 8
-    }
-  }),
-  hamburger: css({
-    '@media print': {
-      display: 'none'
-    },
-    position: 'absolute',
-    top: 0,
-    right: 0,
-    display: 'inline-block',
-    height: HEADER_HEIGHT_MOBILE - 2,
-    width: HEADER_HEIGHT_MOBILE - 2 + 1,
-    [mediaQueries.mUp]: {
-      height: HEADER_HEIGHT - 2,
-      width: HEADER_HEIGHT - 2 + 5
-    }
-  }),
-  menuIcons: css({
-    '@media print': {
-      display: 'none'
-    },
-    position: 'absolute',
-    overflow: 'hidden',
-    top: 0,
-    zIndex: 1,
-    right: HEADER_HEIGHT_MOBILE - 10 + 2,
-    height: HEADER_HEIGHT_MOBILE - 2,
-    [mediaQueries.mUp]: {
-      right: HEADER_HEIGHT - 12 + 5,
-      height: HEADER_HEIGHT - 2
-    }
-  }),
-  search: css({
-    '@media (max-width: 339px)': {
-      display: 'none !important'
-    }
-  }),
-  secondary: css({
-    position: 'absolute',
-    top: 0,
-    left: 15,
-    display: 'inline-block',
-    height: HEADER_HEIGHT_MOBILE,
-    right: `${HEADER_HEIGHT_MOBILE + HEADER_ICON_SIZE}px`,
-    paddingTop: '10px',
-    [mediaQueries.mUp]: {
-      height: HEADER_HEIGHT,
-      right: `${HEADER_HEIGHT + HEADER_HEIGHT}px`,
-      paddingTop: '18px'
-    },
-    transition: `opacity ${TRANSITION_MS}ms ease-in-out`
-  }),
-  sticky: css({
-    position: 'sticky'
-  }),
-  stickyWithFallback: css({
-    // auto prefix does not with multiple values :(
-    // - -webkit-sticky would be missing if not defined explicitly
-    // - glamor 2.20.40 / inline-style-prefixer 3.0.8
-    position: ['fixed', '-webkit-sticky', 'sticky']
-    // - this will produce three position statements
-    // { position: fixed; position: -webkit-sticky; position: sticky; }
-  }),
-  hr: css({
-    margin: 0,
-    display: 'block',
-    border: 0,
-    width: '100%',
-    zIndex: ZINDEX_HEADER
-  }),
-  hrThin: css({
-    height: 1,
-    top: HEADER_HEIGHT_MOBILE - 1,
-    [mediaQueries.mUp]: {
-      top: HEADER_HEIGHT - 1
-    }
-  }),
-  hrThick: css({
-    height: 3,
-    top: HEADER_HEIGHT_MOBILE - 3,
-    [mediaQueries.mUp]: {
-      top: HEADER_HEIGHT - 3
-    }
-  })
-}
-
-// Workaround for WKWebView fixed 0 rendering hickup
-// - iOS 11.4: header is transparent and only appears after triggering a render by scrolling down enough
-const forceRefRedraw = ref => {
-  if (ref) {
-    const redraw = () => {
-      const display = ref.style.display
-      // offsetHeight
-      ref.style.display = 'none'
-      /* eslint-disable-next-line no-unused-expressions */
-      ref.offsetHeight // this force webkit to flush styles (render them)
-      ref.style.display = display
-    }
-    const msPerFrame = 1000 / 30 // assuming 30 fps
-    const frames = [1, 10, 20, 30]
-    // force a redraw on frame x after initial dom mount
-    frames.forEach(frame => {
-      setTimeout(redraw, msPerFrame * frame)
-    })
-  }
-}
-
-const isActiveRoute = (active, route, params = {}) =>
-  !!active &&
-  active.route === route &&
-  Object.keys(params).every(key => active.params[key] === params[key])
-
-const isFront = router => {
-  const active = matchPath(router.asPath)
-
-  return isActiveRoute(active, 'index', {})
-}
-
-const hasBackButton = props =>
-  props.inNativeIOSApp && props.me && !isFront(props.router)
 
 let routeChangeStarted
 
-class Header extends Component {
-  constructor(props) {
-    super(props)
+const Header = ({
+  inNativeApp,
+  inNativeIOSApp,
+  dark: darkProp,
+  me,
+  t,
+  secondaryNav,
+  showSecondary,
+  router,
+  formatColor,
+  pullable = true,
+  hasOverviewNav,
+  children,
+  cover
+}) => {
+  const [isAnyNavExpanded, setIsAnyNavExpanded] = useState(false)
+  const [expandedNav, setExpandedNav] = useState(null)
+  const [isMobile, setIsMobile] = useState()
+  const [headerHeightState, setHeaderHeightState] = useState()
+  const [headerOffset, setHeaderOffset] = useState(0)
+  const [userNavExpanded, setUserNavExpanded] = useState(false)
 
-    this.state = {
-      opaque: !props.cover,
-      mobile: false,
-      expanded: false,
-      backButton: hasBackButton(props),
-      renderSecondaryNav: props.showSecondary
+  const fixedRef = useRef()
+  const diff = useRef(0)
+  const lastY = useRef()
+  const headerHeight = useRef()
+  const lastDiff = useRef()
+
+  const dark = darkProp && !isAnyNavExpanded
+  const textFill = dark ? colors.negative.text : colors.text
+  const logoFill = dark ? colors.logoDark || '#fff' : colors.logo || '#000'
+  const backButton = !hasOverviewNav && inNativeIOSApp && me
+
+  const toggleExpanded = target => {
+    if (target === expandedNav) {
+      setIsAnyNavExpanded(false)
+      setExpandedNav(null)
+    } else if (isAnyNavExpanded) {
+      setExpandedNav(target)
+    } else {
+      setIsAnyNavExpanded(!isAnyNavExpanded)
+      setExpandedNav(target)
     }
+  }
 
-    this.onScroll = () => {
-      const y = window.pageYOffset
+  const openUserNavOverMainNav = () => {
+    setUserNavExpanded(true)
+    setTimeout(() => {
+      setExpandedNav('user')
+    }, TRANSITION_MS)
+  }
 
-      const yOpaque = this.state.mobile ? 70 : 150
-      const opaque = y > yOpaque || !this.props.cover
-      if (opaque !== this.state.opaque) {
-        this.setState(() => ({ opaque }))
+  const closeHandler = () => {
+    if (isAnyNavExpanded) {
+      setIsAnyNavExpanded(false)
+      setExpandedNav(null)
+      setUserNavExpanded(false)
+    }
+  }
+
+  const goTo = (pathName, route) => e => {
+    if (shouldIgnoreClick(e)) {
+      return
+    }
+    e.preventDefault()
+    if (router.pathname === pathName) {
+      window.scrollTo(0, 0)
+      closeHandler()
+    } else {
+      Router.pushRoute(route).then(() => window.scrollTo(0, 0))
+    }
+  }
+
+  useEffect(() => {
+    const onScroll = () => {
+      const y = Math.max(window.pageYOffset, 0)
+
+      if (isAnyNavExpanded) {
+        diff.current = 0
+      } else {
+        const newDiff = lastY.current ? lastY.current - y : 0
+        diff.current += newDiff
+        diff.current = Math.min(
+          Math.max(-headerHeight.current, diff.current),
+          0
+        )
       }
+
+      if (diff.current !== lastDiff.current) {
+        fixedRef.current.style.top = `${diff.current}px`
+        setHeaderOffset(diff.current)
+      }
+
+      lastY.current = y
+      lastDiff.current = diff.current
     }
 
-    this.measure = () => {
+    const measure = () => {
       const mobile = window.innerWidth < mediaQueries.mBreakPoint
-      if (mobile !== this.state.mobile) {
-        this.setState(() => ({ mobile }))
+      if (mobile !== isMobile) {
+        setIsMobile(mobile)
       }
-      this.onScroll()
-    }
-
-    this.close = () => {
-      this.setState({ expanded: false })
-    }
-  }
-
-  componentDidMount() {
-    window.addEventListener('scroll', this.onScroll)
-    window.addEventListener('resize', this.measure)
-    this.measure()
-  }
-
-  componentDidUpdate() {
-    this.measure()
-  }
-
-  componentWillUnmount() {
-    window.removeEventListener('scroll', this.onScroll)
-    window.removeEventListener('resize', this.measure)
-  }
-
-  UNSAFE_componentWillReceiveProps(nextProps) {
-    const backButton = hasBackButton(nextProps)
-    if (this.state.backButton !== backButton) {
-      this.setState({
-        backButton
-      })
-    }
-    clearTimeout(this.secondaryNavTimeout)
-    if (this.state.renderSecondaryNav !== nextProps.showSecondary) {
-      if (nextProps.showSecondary) {
-        this.setState({ renderSecondaryNav: true })
-      } else {
-        this.secondaryNavTimeout = setTimeout(() => {
-          this.setState({ renderSecondaryNav: false })
-        }, TRANSITION_MS)
+      const { height } = fixedRef.current.getBoundingClientRect()
+      headerHeight.current = height
+      if (height !== headerHeightState) {
+        setHeaderHeightState(height)
       }
+      onScroll()
     }
-  }
 
-  render() {
-    const {
-      router,
-      t,
-      me,
-      cover,
-      secondaryNav,
-      showSecondary,
-      formatColor,
-      inNativeApp,
-      inNativeIOSApp,
-      isMember,
-      pullable = true,
-      children
-    } = this.props
-    const { backButton, renderSecondaryNav } = this.state
-
-    const expanded = this.state.expanded
-    const secondaryVisible = showSecondary && !expanded
-    const dark = this.props.dark && !expanded
-
-    const opaque = this.state.opaque || expanded
-    const barStyle = opaque ? merge(styles.bar, styles.barOpaque) : styles.bar
-
-    const bgStyle = opaque
-      ? {
-          backgroundColor: dark ? colors.negative.primaryBg : '#fff'
-        }
-      : undefined
-    const hrColor = dark ? colors.negative.containerBg : colors.divider
-    const hrColorStyle = {
-      color: hrColor,
-      backgroundColor: hrColor
+    window.addEventListener('scroll', onScroll)
+    window.addEventListener('resize', measure)
+    measure()
+    return () => {
+      window.removeEventListener('scroll', onScroll)
+      window.removeEventListener('resize', measure)
     }
-    const textFill = dark ? colors.negative.text : colors.text
-    const logoFill = dark ? colors.logoDark || '#fff' : colors.logo || '#000'
+  }, [isAnyNavExpanded, headerHeightState, isMobile])
 
-    const toggleExpanded = () => {
-      this.setState({ expanded: !expanded })
-      this.props.onNavExpanded && this.props.onNavExpanded(!expanded)
-    }
-    const closeHandler = () => {
-      if (expanded) {
-        toggleExpanded()
+  const headerConfig = useMemo(() => {
+    return [
+      {
+        minWidth: 0,
+        headerHeight:
+          HEADER_HEIGHT_MOBILE +
+          (hasOverviewNav || secondaryNav ? SUBHEADER_HEIGHT : 0) +
+          headerOffset
+      },
+      {
+        minWidth: mediaQueries.mBreakPoint,
+        headerHeight:
+          HEADER_HEIGHT +
+          (hasOverviewNav || secondaryNav ? SUBHEADER_HEIGHT : 0) +
+          headerOffset
       }
-    }
+    ]
+  }, [hasOverviewNav, secondaryNav, headerOffset])
 
-    const goTo = (pathName, route) => e => {
-      if (shouldIgnoreClick(e)) {
-        return
-      }
-      e.preventDefault()
-      if (router.pathname === pathName) {
-        window.scrollTo(0, 0)
-        closeHandler()
-      } else {
-        Router.pushRoute(route).then(() => window.scrollTo(0, 0))
-      }
-    }
-
-    return (
-      <HeaderHeightProvider config={HEADER_HEIGHT_CONFIG}>
-        <ColorContext.Provider value={dark ? colors.negative : colors}>
-          <div
-            {...barStyle}
-            ref={inNativeIOSApp ? forceRefRedraw : undefined}
-            style={bgStyle}
-          >
-            {opaque && (
-              <Fragment>
-                <div
-                  {...styles.center}
-                  style={{ opacity: secondaryVisible ? 0 : 1 }}
-                >
+  return (
+    <HeaderHeightProvider config={headerConfig}>
+      <ColorContext.Provider value={dark ? colors.negative : colors}>
+        <div
+          {...styles.navBar}
+          style={{
+            backgroundColor: dark ? colors.negative.primaryBg : '#fff'
+          }}
+          ref={fixedRef}
+        >
+          <div {...styles.primary}>
+            <div {...styles.navBarItem}>
+              <div {...styles.leftBarItem}>
+                {backButton && (
                   <a
-                    {...styles.logo}
-                    aria-label={t('header/logo/magazine/aria')}
-                    href={'/'}
-                    onClick={goTo('/', 'index')}
-                  >
-                    <Logo fill={logoFill} />
-                  </a>
-                </div>
-                <div
-                  {...styles.leftItem}
-                  style={{
-                    opacity: secondaryVisible || backButton ? 0 : 1
-                  }}
-                >
-                  <User
-                    dark={dark}
-                    me={me}
-                    expanded={expanded}
-                    title={t(`header/nav/${expanded ? 'close' : 'open'}/aria`)}
-                    onClick={toggleExpanded}
-                  />
-                </div>
-                {(inNativeIOSApp || backButton) && (
-                  <a
+                    {...styles.back}
                     style={{
-                      opacity: backButton ? 1 : 0,
+                      opacity: 1,
                       pointerEvents: backButton ? undefined : 'none',
                       href: '#back'
                     }}
@@ -409,97 +216,116 @@ class Header extends Component {
                         }, 200)
                       }
                     }}
-                    {...styles.leftItem}
-                    {...styles.back}
                   >
-                    <BackIcon size={25} fill={textFill} />
+                    <BackIcon size={24} fill={textFill} />
                   </a>
                 )}
-                {secondaryNav && (
-                  <div
-                    {...styles.secondary}
-                    style={{
-                      left: backButton ? 40 : undefined,
-                      opacity: secondaryVisible ? 1 : 0,
-                      pointerEvents: secondaryVisible ? undefined : 'none'
-                    }}
-                  >
-                    {renderSecondaryNav && secondaryNav}
-                  </div>
-                )}
-                <div {...styles.menuIcons}>
-                  {me && <NotificationIcon fill={textFill} />}
-                  {isMember && (
-                    <HeaderIconA
-                      {...styles.search}
-                      title={t('header/nav/search/aria')}
-                      href='/suche'
-                      onClick={goTo('/search', 'search')}
-                    >
-                      <MdSearch fill={textFill} size={HEADER_ICON_SIZE} />
-                    </HeaderIconA>
+                <User
+                  dark={dark}
+                  me={me}
+                  backButton={backButton}
+                  isMobile={isMobile}
+                  id='user'
+                  title={t(
+                    `header/nav/${
+                      expandedNav === 'user' ? 'close' : 'open'
+                    }/aria`
                   )}
-                </div>
-                <div {...styles.hamburger} style={bgStyle}>
-                  <Toggle
-                    dark={dark}
-                    expanded={expanded}
-                    id='primary-menu'
-                    title={t(`header/nav/${expanded ? 'close' : 'open'}/aria`)}
-                    onClick={toggleExpanded}
-                  />
-                </div>
-              </Fragment>
-            )}
+                  onClick={() =>
+                    !isAnyNavExpanded
+                      ? toggleExpanded('user')
+                      : expandedNav !== 'user'
+                      ? openUserNavOverMainNav()
+                      : closeHandler()
+                  }
+                />
+                {me && <NotificationIcon fill={textFill} />}
+              </div>
+            </div>
+            <div {...styles.navBarItem}>
+              <a
+                {...styles.logo}
+                aria-label={t('header/logo/magazine/aria')}
+                href={'/'}
+                onClick={goTo('/', 'index')}
+              >
+                <Logo fill={logoFill} />
+              </a>
+            </div>
+            <div {...styles.navBarItem}>
+              <div {...styles.rightBarItem}>
+                <Toggle
+                  expanded={isAnyNavExpanded}
+                  dark={dark}
+                  size={28}
+                  title={t(
+                    `header/nav/${
+                      expandedNav === 'main' ? 'close' : 'open'
+                    }/aria`
+                  )}
+                  id='main'
+                  onClick={() =>
+                    isAnyNavExpanded ? closeHandler() : toggleExpanded('main')
+                  }
+                />
+              </div>
+            </div>
           </div>
-          {opaque && (
-            <hr
-              {...styles.stickyWithFallback}
-              {...styles.hr}
-              {...styles[formatColor ? 'hrThick' : 'hrThin']}
-              style={
-                formatColor
-                  ? {
-                      color: formatColor,
-                      backgroundColor: formatColor
-                    }
-                  : hrColorStyle
+          <SecondaryNav
+            secondaryNav={secondaryNav}
+            router={router}
+            dark={dark}
+            formatColor={formatColor}
+            showSecondary={showSecondary}
+            hasOverviewNav={hasOverviewNav}
+            isSecondarySticky={headerOffset === -headerHeightState}
+          />
+          <HLine formatColor={formatColor} dark={dark} />
+        </div>
+        <Popover formatColor={formatColor} expanded={expandedNav === 'main'}>
+          <NavPopover
+            me={me}
+            router={router}
+            expanded={expandedNav === 'main'}
+            closeHandler={closeHandler}
+            onSearchSubmit={closeHandler}
+          />
+        </Popover>
+        <Popover
+          formatColor={formatColor}
+          expanded={userNavExpanded || expandedNav === 'user'}
+        >
+          <UserNavPopover
+            me={me}
+            router={router}
+            expanded={userNavExpanded || expandedNav === 'user'}
+            closeHandler={closeHandler}
+          />
+        </Popover>
+        <LoadingBar
+          onRouteChangeStart={() => {
+            routeChangeStarted = true
+          }}
+        />
+        {inNativeApp && pullable && (
+          <Pullable
+            dark={dark}
+            onRefresh={() => {
+              if (inNativeIOSApp) {
+                postMessage({ type: 'haptic', payload: { type: 'impact' } })
               }
-            />
-          )}
-          <Popover expanded={expanded}>
-            <NavPopover
-              me={me}
-              router={router}
-              expanded={expanded}
-              closeHandler={closeHandler}
-            />
-          </Popover>
-          <LoadingBar
-            onRouteChangeStart={() => {
-              routeChangeStarted = true
+              // give the browser 3 frames (1000/30fps) to start animating the spinner
+              setTimeout(() => {
+                window.location.reload(true)
+              }, 33 * 3)
             }}
           />
-          {!!cover && <div {...styles.cover}>{cover}</div>}
-          {inNativeApp && pullable && (
-            <Pullable
-              dark={dark}
-              onRefresh={() => {
-                if (inNativeIOSApp) {
-                  postMessage({ type: 'haptic', payload: { type: 'impact' } })
-                }
-                // give the browser 3 frames (1000/30fps) to start animating the spinner
-                setTimeout(() => {
-                  window.location.reload(true)
-                }, 33 * 3)
-              }}
-            />
-          )}
-        </ColorContext.Provider>
-        {children}
-      </HeaderHeightProvider>
-    )
-  }
+        )}
+      </ColorContext.Provider>
+      {cover}
+      {children}
+    </HeaderHeightProvider>
+  )
 }
 
 export default compose(
@@ -508,3 +334,65 @@ export default compose(
   withRouter,
   withInNativeApp
 )(Header)
+
+const styles = {
+  navBar: css({
+    height: HEADER_HEIGHT_MOBILE,
+    zIndex: ZINDEX_POPOVER + 1,
+    position: 'fixed',
+    top: 0,
+    left: 0,
+    right: 0,
+    [mediaQueries.mUp]: {
+      height: HEADER_HEIGHT
+    },
+    '@media print': {
+      position: 'absolute'
+    }
+  }),
+  primary: css({
+    display: 'flex',
+    justifyContent: 'space-between'
+  }),
+  navBarItem: css({
+    flex: 1,
+    display: 'flex',
+    justifyContent: 'center'
+  }),
+  leftBarItem: css({
+    marginRight: 'auto',
+    display: 'flex',
+    justifyContent: 'flex-start',
+    width: '100%'
+  }),
+  rightBarItem: css({
+    marginLeft: 'auto',
+    width: '100%',
+    display: 'flex',
+    justifyContent: 'flex-end',
+    height: HEADER_HEIGHT_MOBILE - 2,
+    [mediaQueries.mUp]: {
+      height: HEADER_HEIGHT - 2
+    },
+    '@media print': {
+      display: 'none'
+    }
+  }),
+  back: css({
+    display: 'block',
+    padding: '12px 0px 12px 12px',
+    [mediaQueries.mUp]: {
+      top: -1 + 8
+    }
+  }),
+  logo: css({
+    display: 'block',
+    padding: LOGO_PADDING_MOBILE,
+    width: LOGO_WIDTH_MOBILE + LOGO_PADDING_MOBILE * 2,
+    verticalAlign: 'middle',
+    [mediaQueries.mUp]: {
+      padding: LOGO_PADDING,
+      width: LOGO_WIDTH + LOGO_PADDING * 2
+    }
+  })
+}
